chore: made RemoteImage a tsx file and simplified some logic
